fix(safety): final-review fixes — lot rounding blocker, floor snapshot, safety_floor.yaml consumption, manager reason padding
Final whole-branch review verdict was NO-GO with 1 BLOCKING + 3 IMPORTANT:

- [BLOCKING] _units_to_lots now floors to lot_step (never half-rounds up)
  and REJECTS below-broker-minimum sizes with MT5Error instead of
  clamping up to volume_min (which meant ~18x effective leverage vs the
  5x cap on a Standard account at R1000). Live-cutover runbook now
  mandates Exness Standard Cent and documents why.
- [IMPORTANT] tb status floor: RatchetFloor.current_floor is a @property;
  calling it raised a swallowed TypeError leaving floor null forever —
  broke the runbook's floor verification. Fixed + snapshot test.
- [IMPORTANT] config/safety_floor.yaml is now actually consumed:
  load_config() overlays it floor-wins onto settings.yaml with the two
  key renames the risk modules read (daily_drawdown_stop_pct ->
  daily_drawdown_limit_pct, max_leverage_effective ->
  max_effective_leverage); circuit breaker api_error_limit now reads
  circuit_breaker.api_error_threshold (10 per the safety file, was
  hardcoded 5).
- [IMPORTANT] manager tune reasons shorter than the queue's
  MIN_REASON_LEN are padded in _enqueue_tune so a validated tune can't
  be rejected after manager_log recorded it applied (audit divergence).

Tests: 10 new across mt5 orders / control queue / safety floor / manager
runner. Full suite 395 passed.

+++ src/control/queue.py
@@ -562,7 +562,8 @@ class ControlQueue:
             if self.risk_manager is not None:
                 floor = getattr(self.risk_manager, "ratchet_floor", None)
                 if floor is not None and hasattr(floor, "current_floor"):
-                    snapshot["floor"] = floor.current_floor()
+                    # current_floor is a @property — do not call it.
+                    snapshot["floor"] = float(floor.current_floor)
         except Exception as e:
             logger.debug(f"status_snapshot: floor unavailable: {e}")
